requirements.txt
streamlit==1.28.1
openai==1.3.0
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3
pandas==2.1.1
plotly==5.15.0
